Changed a few things in relation to fonts and background
I made the descision on what font I am using for my headers and main
text.

I also removed the coffee3.jpg file from the #hero div and made it part
of the actual body. Made it static and also reduced the resolution of
the image to improve load times.